# backend/src/main/java/com/platform/user/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: User Module RBAC API
  description: |
    Comprehensive Role-Based Access Control (RBAC) API for the Spring Boot Modulith payment platform.

    This API provides multi-tenant permission management with organization isolation, custom role creation,
    and real-time permission checking with sub-200ms performance requirements.

    ## Features
    - Multi-tenant organization isolation
    - System and custom role management
    - Resource×Action permission model
    - Role assignment with expiration support
    - Comprehensive caching for performance
    - Event-driven architecture

    ## Authentication
    All endpoints require valid JWT authentication with appropriate permissions.

    ## Rate Limiting
    - Permission checks: 1000 requests/minute per user
    - Role management: 100 requests/minute per user
    - User assignments: 500 requests/minute per user
  version: 1.0.0
  contact:
    name: Platform Team
    email: platform@company.com
  license:
    name: Proprietary
    url: https://company.com/license
servers:
  - url: https://api.platform.com/v1
    description: Production server
  - url: https://staging-api.platform.com/v1
    description: Staging server
  - url: http://localhost:8080
    description: Local development server
security:
  - BearerAuth: []
...
paths:
...
  /api/organizations/{organizationId}/roles:
    get:
      tags: [Roles]
      summary: Get organization roles
      description: Get all roles for an organization (predefined and custom)
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: List of organization roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolesResponse'
    post:
      tags: [Roles]
      summary: Create custom role
      description: |
        Create a new custom role in an organization. Only organization admins
        can create custom roles.
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleCreatedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Role name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/organizations/{organizationId}/roles/{roleId}:
    get:
      tags: [Roles]
      summary: Get role details
      description: Get detailed information about a specific role including permissions
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/RoleId'
      responses:
        '200':
          description: Role details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDetailsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Roles]
      summary: Update custom role
      description: Update an existing custom role (predefined roles cannot be modified)
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/RoleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleUpdatedResponse'
    delete:
      tags: [Roles]
      summary: Delete custom role
      description: |
        Delete a custom role. Only custom roles with no active assignments
        can be deleted.
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/RoleId'
      responses:
        '200':
          description: Role deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDeletedResponse'
...
  /api/organizations/{organizationId}/users/{userId}/roles/{roleId}:
    delete:
      tags: [User Roles]
      summary: Remove role from user
      description: Remove a role assignment from a user with audit tracking
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RoleId'
      responses:
        '200':
          description: Role removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleRemovalResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        example: 1
      description: Organization identifier
    RoleId:
      name: roleId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        example: 5
      description: Role identifier
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        example: 10
      description: User identifier
  schemas:
    PermissionInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        resource:
          type: string
          example: PAYMENTS
        action:
          type: string
          example: WRITE
        description:
          type: string
          example: "Write access to payment operations"
      required:
        - id
        - resource
        - action
...
    RoleInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 5
        name:
          type: string
          example: payment-manager
        description:
          type: string
          example: "Manages payment processing and customer billing"
        roleType:
          type: string
          enum: [PREDEFINED, CUSTOM]
          example: CUSTOM
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        assignedUserCount:
          type: integer
          format: int64
          example: 3
    RoleDetails:
      allOf:
        - $ref: '#/components/schemas/RoleInfo'
        - type: object
          properties:
            createdBy:
              type: integer
              format: int64
              example: 1
            permissions:
              type: array
              items:
                $ref: '#/components/schemas/PermissionInfo'
    RoleStatistics:
      type: object
      properties:
        totalRoles:
          type: integer
          format: int64
          example: 8
        customRoles:
          type: integer
          format: int64
          example: 4
        predefinedRoles:
          type: integer
          format: int64
          example: 4
        totalAssignedUsers:
          type: integer
          format: int64
          example: 25
    RolesResponse:
      type: object
      properties:
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleInfo'
        statistics:
          $ref: '#/components/schemas/RoleStatistics'
    RoleDetailsResponse:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/RoleDetails'
    CreateRoleRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          example: payment-manager
        description:
          type: string
          maxLength: 500
          example: "Manages payment processing and customer billing"
        permissionIds:
          type: array
          items:
            type: integer
            format: int64
          minItems: 1
          example: [1, 2, 5]
      required:
        - name
        - permissionIds
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          example: payment-specialist
        description:
          type: string
          maxLength: 500
          example: "Specialized payment processing role"
        permissionIds:
          type: array
          items:
            type: integer
            format: int64
          example: [1, 2, 5, 8]
      required:
        - name
    RoleCreatedResponse:
      type: object
      properties:
        status:
          type: string
          example: Success
        roleName:
          type: string
          example: payment-manager
        message:
          type: string
          example: "Role created with 3 permissions"
    RoleUpdatedResponse:
      type: object
      properties:
        status:
          type: string
          example: Success
        message:
          type: string
          example: "Role updated successfully"
    RoleDeletedResponse:
      type: object
      properties:
        status:
          type: string
          example: Success
        message:
          type: string
          example: "Role deleted successfully"
...
    RoleRemovalResponse:
      type: object
      properties:
        status:
          type: string
          example: Success
        message:
          type: string
          example: "Role removed successfully"
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid request parameters"
        correlationId:
          type: string
          example: "abc123-def456-ghi789"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'